Ask for user input and return the coordinates of a valid move on board
Coordinates are specified as a letter and a number, for example, A2 is the first row of the second column, and C1 is the third row of the first column.

The function returns an array of integers containing the row and column number corresponding to the input.

The function returns an array of integers containing the row and column number corresponding to the input.

The integers indicate a valid (empty) position on the board.

If the user provides coordinates that are outside of a 26x26 sized board, keep asking.

If the user provides coordinates for a place that is taken, keep asking.

If the user provides coordinates for a place that is taken, keep asking.

If the user provides input that does not follow the format of coordinates, keep asking.